% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoRateModel.R
\name{fitTwoRateReachModel}
\alias{fitTwoRateReachModel}
\title{Fit the Two-Rate Model To a Dataset.}
\usage{
fitTwoRateReachModel(reaches, rotations, fitInitialState = FALSE,
  oneTwoRates = 2, verbose = FALSE)
}
\arguments{
\item{reaches}{A sequence of reach deviations.}

\item{rotations}{A sequence of feedback manipulations.}

\item{fitInitialState}{Boolean: is the initial state zero or does it have to be fit.}

\item{oneTwoRates}{How many processes to fit? (1 or 2)}

\item{verbose}{Should detailed information be outputted during the fitting?}
}
\value{
The set of parameters that minimizes the difference between model output and the \code{reaches} given the \code{rotations}.
}
\description{
Fit the Two-Rate Model To a Dataset.
}
