% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RotAdapt}
\alias{RotAdapt}
\title{Rotation Adaptation data for use with the two-rate model}
\format{A data frame with 288 rows and 2 variables:
\describe{
  \item{schedule}{perturbation schedule, for use in the two-rate model}
  \item{reaches}{the median initial reach errors across a group of participants}
}}
\usage{
RotAdapt
}
\description{
A dataset with a perturbation schedule and the reach direction errors of a group
of participants perofrming an experiment with that schedule.
}
\keyword{datasets}
