% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{twoRateReachModel}
\alias{twoRateReachModel}
\title{Evaluate the Two-Rate Model.}
\usage{
twoRateReachModel(par, schedule)
}
\arguments{
\item{par}{A list of named parameters.}

\item{schedule}{A sequence of feedback manipulations}
}
\description{
Evaluate the Two-Rate Model Given Parameters and a Perturbation Schedule.
}
\seealso{
\code{\link{fitTwoRateReachModel}} and \code{\link{twoRateReachModelErrors}}
}
