% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{twoRateReachModelErrors}
\alias{twoRateReachModelErrors}
\title{Return the Mean Squared Error Between a Two-Rate Model and a Dataset.}
\usage{
twoRateReachModelErrors(par, reaches, schedule, checkStability = TRUE)
}
\arguments{
\item{par}{A list of named parameters.}

\item{reaches}{A list of reach deviations.}

\item{schedule}{A sequence of feedback manipulations}
}
\value{
The average squared error of the model's prediction of the reach deviations given the rotation schedule and model parameters.
}
\description{
Calculates the average squared error of the Two-Rate Model's prediction, given a set of parameters and a rotation schedule.
}
\seealso{
\code{\link{fitTwoRateReachModel}} and \code{\link{twoRateReachModel}}
}
